moved caching to base so that multiple classes can access - caching of table works
