+++ dataentry/views/base.py
@@ -7,6 +7,8 @@ from api.ehb_service_client import ServiceClient
 from api.utilities import DriverUtils
 from api.models.protocols import ProtocolDataSource
 
+from django.core.cache import cache
+
 log = logging.getLogger(__name__)
 
 
@@ -79,3 +81,20 @@ class DataEntryView(TemplateView):
             context['label_id'] = self.request.GET.get('label_id', 1)
         context['label'] = self.get_label(context)
         return context
+
+    def check_cache(self, cache_key):
+        self.cached_data = cache.get(cache_key)
+        if self.cached_data:
+            return True
+        else:
+            return False
+
+    def get_cache (self, cache_key):
+        print ("we're in get cache")
+        print (cache.get(cache_key))
+        return cache.get(cache_key)
+
+    def create_cache(self, cache_key, cache_data):
+        cache.set(cache_key, cache_data)
+        cache.persist(cache_key)
+        self.check_cache(cache_key)
